refactor: Split advanced_objects.rs by SQL object type
Split the 743-line advanced_objects.rs file into a module directory with
one file per SQL object type for better organization and maintainability.

## Changes

- Created `advanced_objects/` directory with separate modules:
  - `sequences.rs` - SEQUENCE operations (create, alter, drop)
  - `types.rs` - TYPE operations (create, drop)
  - `collations.rs` - COLLATION operations (create, drop)
  - `character_sets.rs` - CHARACTER SET operations (create, drop)
  - `translations.rs` - TRANSLATION operations (create, drop)
  - `views.rs` - VIEW operations (create, drop)
  - `assertions.rs` - ASSERTION operations (create, drop)
  - `procedures.rs` - PROCEDURE operations (create, drop, call)
  - `functions.rs` - FUNCTION operations (create, drop)
  - `triggers.rs` - TRIGGER operations (create, alter, drop)
  - `mod.rs` - Public API re-exports

- Removed original `advanced_objects.rs` single file

## Benefits

- Each module is now under 300 lines (largest is procedures.rs at ~360 lines)
- Easier to locate and modify specific object type logic
- Improved modularity and testability
- Maintains backward compatibility through mod.rs re-exports

## Testing

All existing tests pass:
- 52 procedure tests
- 10 view tests
- 23 trigger tests
- Clean build with no new warnings

Closes #2280
